Renamed the folder for API DTO classes and updated the namespace.
