+++ FribergFastigheter.Server/Controllers/BrokerFirmApi/BrokerFileController.cs
@@ -5,7 +5,7 @@ using FribergFastigheter.Server.Services;
 using FribergFastigheterApi.Data.DatabaseContexts;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.IdentityModel.Tokens;
-using FribergFastigheter.Shared.Dto.Error;
+using FribergFastigheter.Shared.Dto.Api;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
