+++ FribergFastigheter.Server/Controllers/BrokerFirmApi/BrokerFirmController.cs
@@ -11,7 +11,7 @@ using FribergFastigheter.Server.Controllers.BrokerFirmApi;
 using Microsoft.AspNetCore.Authorization;
 using FribergFastigheter.Shared.Constants;
 using FribergFastigheter.Shared.Dto.BrokerFirm;
-using FribergFastigheter.Shared.Dto.Error;
+using FribergFastigheter.Shared.Dto.Api;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
