+++ FribergFastigheter.Server/Controllers/BrokerFirmApi/BrokerHousingController.cs
@@ -9,9 +9,9 @@ using Microsoft.AspNetCore.Authorization;
 using FribergFastigheter.Shared.Constants;
 using FribergFastigheter.Shared.Dto.Housing;
 using FribergFastigheter.Shared.Dto.Image;
-using FribergFastigheter.Shared.Dto.Error;
 using FribergFastigheter.Shared.Enums;
 using FribergFastigheter.Shared.Services.AuthorizationHandlers.Housing.Data;
+using FribergFastigheter.Shared.Dto.Api;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
